chore: move chyt api [YTFRONT-4219]
